+++ CHANGELOG.md
@@ -6,7 +6,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## \[Unreleased\]
-
 ### New features
 - Add CocoRoboflowImporter
   (<https://github.com/openvinotoolkit/datumaro/pull/976>)
@@ -14,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Enhancements
 - Use autosummary for fully-automatic Python module docs generation
   (<https://github.com/openvinotoolkit/datumaro/pull/973>)
+
+### Bug fixes
+- Fix Mapillary Vistas data format (<https://github.com/openvinotoolkit/datumaro/pull/977>)
 
 ## 20/04/2023 - Release 1.2.0
 ### New features